// admin/frontend/src/services/api/admin/types.ts
export interface NotionConfig {
  clientId: string;
  clientSecret: string;
  redirectUri: string;
  syncButtonEnabled: boolean;
}